use context/provider for better component composition
